requirements.txt
PyYAML>=5.4.1
requests-random-user-agent>=2021.5.29
